<refactoring> : 다리 생성 테스트 코드 리팩토링 기존에 ApplicationTest에 작성되어 있었던 다리 생성 관련 테스트 코드로 수정함.
