+++ src/test/java/bridge/ApplicationTest.java
@@ -13,16 +13,7 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 class ApplicationTest extends NsTest {
-
     private static final String ERROR_MESSAGE = "[ERROR]";
-
-    @Test
-    void 다리_생성_테스트() {
-        BridgeNumberGenerator numberGenerator = new TestNumberGenerator(newArrayList(1, 0, 0));
-        BridgeMaker bridgeMaker = new BridgeMaker(numberGenerator);
-        List<String> bridge = bridgeMaker.makeBridge(3);
-        assertThat(bridge).containsExactly("U", "D", "D");
-    }
 
     @Test
     void 기능_테스트() {
@@ -53,19 +44,5 @@ class ApplicationTest extends NsTest {
     @Override
     protected void runMain() {
         Application.main(new String[]{});
-    }
-
-    static class TestNumberGenerator implements BridgeNumberGenerator {
-
-        private final List<Integer> numbers;
-
-        TestNumberGenerator(List<Integer> numbers) {
-            this.numbers = numbers;
-        }
-
-        @Override
-        public int generate() {
-            return numbers.remove(0);
-        }
     }
 }
